requirements.txt
flask==3.0.0
gunicorn==21.2.0
google-generativeai==0.8.0
google-cloud-storage==2.14.0
google-cloud-aiplatform==1.38.0
chromadb==0.5.0
langchain-text-splitters==0.2.4
PyPDF2==3.0.1
python-dotenv==1.0.0
numpy<2.0.0

